[feat] Material3 마이그레이션 - SemesterListScreen, SemesterDetailScreen (#53)
* feat: SemesterListScreen ui refactor

* feat: SemesterDetailScreen ui refactor

* feat: SemesterDetailScreen ui refactor

* feat: enableEdgeToEdge on MainActivity

* chore: change maxValue's type to nullable

* chore: delete initial LaunchedEffect
