+++ app/src/main/kotlin/com/yourssu/soomsil/usaint/MainActivity.kt
@@ -3,12 +3,12 @@ package com.yourssu.soomsil.usaint
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
 import androidx.activity.viewModels
 import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.rememberNavController
-import com.yourssu.design.system.compose.YdsTheme
 import com.yourssu.soomsil.usaint.screen.USaintNavHost
 import com.yourssu.soomsil.usaint.screen.home.navigation.Home
 import com.yourssu.soomsil.usaint.screen.login.navigation.Login
@@ -21,7 +21,7 @@ class MainActivity : ComponentActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-//        enableEdgeToEdge()
+        enableEdgeToEdge()
 
         setContent {
             SoomsilUSaintTheme {
